code smell/duplication
